feature(theme): Delete all custom css and template overrides
We'll use Aalborg theme instead, which comes with Elgg 1.9.
It is both prettier and mobile-friendly. Big wins.

The AJAX features were not hardened enough to be reliable. Too many bugs.
Also, some basic things like likes and friending are ajaxed by default now.
Elgg should get better and better at this over time so there's not a huge
need for us to do tons of custom stuff that might be subtly upgrade-incompatible.